debugged lambda

+++ src/global_variables.f90
@@ -6,6 +6,7 @@ module global_variables
     
     integer :: nc,nv,nex,nb,northo,ncutoff,nex_max,biex,num_ortho,nex_max1,nex_max_s,nex_max_t,lin_ind  !no of conduction, valence, exciton states and c+v total no of states 
     integer :: nk,nQ,neb
+    
     double precision, allocatable :: kpts(:,:),Qpts(:,:)
     end type input
     type(input) :: sys
@@ -37,7 +38,7 @@ module global_variables
           double precision, allocatable :: xi_d_in_ttss(:,:,:,:),xi_d_in_sstt(:,:,:,:)
           double precision, allocatable :: xi_x_in_ttss(:,:,:,:),xi_x_in_sstt(:,:,:,:)
           double precision, allocatable :: xi_x_ttss(:,:,:,:),xi_x_sstt(:,:,:,:)
-          double precision,allocatable :: H(:,:),O(:,:),c(:,:)
+          complex(kind=16),allocatable :: H(:,:),O(:,:),c(:,:)
     end  type Hamiltonian
